The virtual machine is built (Mac OSX)
There are problems with instantiations of std::basic_ostringstream<T>, where T is a character type.
The problem seems to come from the implementation provided by libc++ which is the default in Mac.

See http://lists.apple.com/archives/xcode-users/2012/Jul/msg00095.html for a discussion about the
problem and a possible way to achieve the desired functionality. For now I decided to avoid any
use of ostringstream by _breaking_ some string functionality.