Cleanup: removed the need to explictly write Copyable.
The copyable thing is actually an optimization, which the "user"
of the store meta-program (i.e., the VM developer) should not
need care about.

Now the generator infers the copyable property automatically if
and only if the type has a Value behavior *and* is StoredAs<T>
(with a T that can *actually* be stored in a MemWord, without
external memory).