Added a builtin Record.adjoinAtIfHasFeature.
This is partial work of Record.adjoinAt that succeeds if the
original record already has the feature to adjoin. In this case,
a fast copy is done in C++. Otherwise, the function does not
succeed and AdjoinAt should fallback on a slower implementation.

Eventually, this will be replaced by a complete Record.adjoinAt
as a builtin. But we need this partial implementation to be fast
right now so that PEG.oz is not too slow.